@@ -12,16 +12,16 @@
         }
     ],
     "require": {
-        "php": ">=5.5",
+        "php": "^8.1",
         "aws/aws-sdk-php": "^3.2.6",
-        "symfony/config": "~2.3|~3.0|~4.0|~5.0",
-        "symfony/dependency-injection": "~2.3|~3.0|~4.0|~5.0",
-        "symfony/http-kernel": "~2.3|~3.0|~4.0|~5.0"
+        "symfony/config": "~5.0|^6.0",
+        "symfony/dependency-injection": "~5.0|^6.0",
+        "symfony/http-kernel": "~5.0|^6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.35|^5.4.0|^6.0.0",
-        "symfony/framework-bundle": "~2.3|~3.0|~4.0|~5.0",
-        "symfony/yaml": "~2.3|~3.0|~4.0|~5.0"
+        "phpunit/phpunit": "^9.5",
+        "symfony/framework-bundle": "~5.0|^6.0",
+        "symfony/yaml": "~5.0|^6.0"
     },
     "autoload": {
         "psr-4": {
